move offline calculation into its own file
